minor fixes: fixed failing tests after rebase

+++ examples/torch_code/if_else_contexts.py
@@ -37,7 +37,7 @@ def classify(x):
 class PiecewiseNet(nn.Module):
     def __init__(self, threshold):
         super().__init__()
-        self.threshold = nn.Parameter(torch.as_tensor(threshold).float())
+        self.threshold = nn.Parameter(torch.as_tensor(threshold))
         self.learnable_params = [self.threshold]
 
     def forward(self, x):
